build updates  minor compile fixes

+++ common/tnmsCore/include/TnmsOid.h
@@ -63,7 +63,7 @@ public:
     // Serializable
     virtual ssize_t  serialize       ( char * buffer, size_t len );
     virtual ssize_t  deserialize     ( const char * buffer, size_t len );
-    virtual uint32_t size() const;
+    virtual size_t   size() const;
     virtual uint16_t message_type() const { return 0; }
 
 public:
